Modify the directory structure
