+++ src/Page/Home/Components/CommentDrawer/CommentDrawer.tsx
@@ -1,273 +1,14 @@
-import * as Icon from '@Icons/index';
-import * as Layout from '@Layouts/DefaultLayout';
+import * as S from './CommentDrawer.styled';
+import * as C from './CommentDrawer.component';
+
 import * as Dialog from '@radix-ui/react-dialog';
-import { AnimatePresence, useAnimate } from 'framer-motion';
-import * as S from './HomePage.styled';
 
-import {
-  MenuHeadProp,
-  MenuIconProp,
-  MenuItem,
-  MenuItemProp,
-  MenuListProp,
-  SideBarMenuProp,
-} from './HomePage.types';
+import * as Layout from '@Layouts/DefaultLayout';
 
-import { TransitionDirection } from '@Components/ContentCard.types';
-import { useEffect, useRef, useState } from 'react';
-import ContentCard from '@Components/ContentCard';
+import * as Icon from '@Icons/index';
 import InlineProfile from '@Components/InlineProfile';
 
-const sidebarMenuList: MenuItem[] = [
-  {
-    name: '홈',
-    icon: Icon.Home,
-    path: '/home',
-    enabled: true,
-  },
-  {
-    name: '내 피드',
-    icon: Icon.Profile,
-    path: '/profile',
-    enabled: true,
-  },
-  {
-    name: '명예의 전당',
-    icon: Icon.Hamburger,
-    path: '/cool',
-    enabled: false,
-  },
-  {
-    name: '추가 기능',
-    icon: Icon.Hamburger,
-    path: '/addon',
-    enabled: false,
-  },
-];
-
-const MiniLogo = () => (
-  <S.MenuFooter>
-    <Icon.Logo
-      css={{
-        width: '$icon-sm',
-        height: '$icon-sm',
-      }}
-    />
-    <span>컴플리메이트</span>
-  </S.MenuFooter>
-);
-
-const MenuIcon = ({ IconInner, ...rest }: MenuIconProp) => {
-  return (
-    <IconInner
-      css={{
-        width: '$icon-menu',
-        height: '$icon-menu',
-        color: '$body',
-        ...rest,
-      }}
-    />
-  );
-};
-
-const Head = ({ onBackClick, onSettingClick }: MenuHeadProp) => {
-  return (
-    <S.SideBarHeader>
-      <div className="group">
-        <button onClick={onBackClick}>
-          <MenuIcon IconInner={Icon.Left} />
-        </button>
-        <p>메뉴</p>
-      </div>
-
-      <button onClick={onSettingClick}>
-        <Icon.Setting />
-      </button>
-    </S.SideBarHeader>
-  );
-};
-
-const Item = ({ selected, item, onClick }: MenuItemProp) => {
-  return (
-    <li
-      className={!item.enabled ? 'disabled' : selected ? 'selected' : ''}
-      onClick={() => item.enabled && onClick(item.path)}
-    >
-      <MenuIcon
-        IconInner={item.icon}
-        color={!item.enabled ? '$depth3' : selected ? '$point' : ''}
-      />
-      {item.name}
-    </li>
-  );
-};
-
-const List = ({ list, currentPath, onMenuItemClick }: MenuListProp) => {
-  return (
-    <S.MenuList>
-      {list.map((item) => (
-        <Item
-          key={item.path}
-          selected={currentPath === item.path}
-          item={item}
-          onClick={onMenuItemClick}
-        />
-      ))}
-    </S.MenuList>
-  );
-};
-
-const SideMenu = { Head, List };
-
-export const SideBarMenu = ({
-  container,
-  onBackClick,
-  onSettingClick,
-  onMenuItemClick,
-  overlayScope,
-  contentScope,
-}: SideBarMenuProp) => {
-  return (
-    <Dialog.Portal container={container.current}>
-      <S.DialogOverlay ref={overlayScope} initial={{ opacity: 0 }} animate={{ opacity: 1 }}>
-        <S.DialogContent ref={contentScope} initial={{ x: '-100%' }} animate={{ x: 0 }}>
-          <S.MenuInner>
-            <Layout.Head>
-              <SideMenu.Head onBackClick={onBackClick} onSettingClick={onSettingClick} />
-            </Layout.Head>
-
-            <Layout.Body>
-              <SideMenu.List
-                list={sidebarMenuList}
-                currentPath="/home"
-                onMenuItemClick={onMenuItemClick}
-              />
-            </Layout.Body>
-
-            <Layout.Foot>
-              <MiniLogo />
-            </Layout.Foot>
-          </S.MenuInner>
-        </S.DialogContent>
-      </S.DialogOverlay>
-    </Dialog.Portal>
-  );
-};
-
-const contentSectionVariants = {
-  before: (direction: number) => ({ y: `${100 * direction}%` }),
-  normal: { y: 0 },
-  exit: (direction: number) => ({ y: `${-100 * direction}%` }),
-};
-
-export const DraggablePost = ({ post, onTransitionRaise, drawerRef, onCommentClick }) => {
-  const direction = useRef(TransitionDirection.Down);
-
-  const handleTransitionRaise = (newDirection: TransitionDirection) => {
-    direction.current = newDirection;
-    onTransitionRaise(newDirection);
-  };
-
-  return (
-    <AnimatePresence custom={direction.current} initial={false}>
-      <S.ContentSection
-        key={post.id}
-        custom={direction.current}
-        variants={contentSectionVariants}
-        initial="before"
-        animate="normal"
-        exit="exit"
-      >
-        <ContentCard
-          post={post}
-          onTransitionRaise={handleTransitionRaise}
-          drawerRef={drawerRef}
-          onCommentClick={onCommentClick}
-        />
-      </S.ContentSection>
-    </AnimatePresence>
-  );
-};
-
-const CommentItemInner = ({ name, profile, date, body, replys, heartCount }) => {
-  return (
-    <div>
-      <Layout.Head>
-        <div className="left">
-          <InlineProfile type="normal" nickname={name} profile={profile} />
-          <p className="time">{date}</p>
-        </div>
-
-        <button>
-          <Icon.More
-            css={{
-              width: '$icon-comment',
-              height: '$icon-comment',
-            }}
-          />
-        </button>
-      </Layout.Head>
-
-      <Layout.Body>
-        <p className="body">{body}</p>
-      </Layout.Body>
-
-      <Layout.Foot>
-        {replys && (
-          <button>
-            <Icon.Comment
-              css={{
-                width: '$icon-comment',
-                height: '$icon-comment',
-                lineHeight: 0,
-                color: '$point',
-              }}
-            />
-            {replys.length}
-          </button>
-        )}
-        <button>
-          <Icon.Heart
-            css={{
-              width: '$icon-comment',
-              height: '$icon-comment',
-              lineHeight: 0,
-              color: '$point',
-            }}
-          />
-          {heartCount}
-        </button>
-      </Layout.Foot>
-    </div>
-  );
-};
-
-const CommentItem = ({ comment }) => {
-  return (
-    <div>
-      <CommentItemInner {...comment} />
-      {comment.replys.length !== 0 && (
-        <div className="padding">
-          {comment.replys.map((reply) => (
-            <CommentItemInner {...reply} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-};
-
-const CommentList = ({ comments }) => {
-  return (
-    <>
-      {comments.map((comment) => (
-        <CommentItem key={comment.id} comment={comment} />
-      ))}
-    </>
-  );
-};
-
-export const CommentDrawer = ({ comments, container, drawerScope }) => {
+export const CommentDrawer = ({ comments, container, drawerScope }: CommentDrawerProp) => {
   return (
     <Dialog.Portal container={container}>
       <S.DrawerContent ref={drawerScope} initial={{ y: '100%' }} animate={{ y: 0 }}>
@@ -296,7 +37,7 @@ export const CommentDrawer = ({ comments, container, drawerScope }) => {
                 </div>
               </Layout.Head>
 
-              <CommentList comments={comments} />
+              <C.List comments={comments} />
             </Layout.Body>
 
             <Layout.Foot>
