+++ src/Page/Home/Components/ContentCard/ContentCard.types.tsx
