+++ src/Page/Home/Components/SidebarMenu/SidebarMenu.types.tsx
@@ -1,17 +1,7 @@
-import { TransitionDirection } from '@Components/ContentCard.types';
 import { Left as IconComponent } from '@Icons/index';
 import Stitches from '@stitches/react';
-import { AnimationScope } from 'framer-motion';
-import { ComponentType, RefObject } from 'react';
-
-export type SideBarMenuProp = {
-  container: RefObject<HTMLElement>;
-  onBackClick: () => void;
-  onSettingClick: () => void;
-  onMenuItemClick: (path: string) => void;
-  overlayScope: AnimationScope<any>;
-  contentScope: AnimationScope<any>;
-};
+import type { AnimationScope } from 'framer-motion';
+import type { ComponentType, RefObject } from 'react';
 
 export type MenuItem = {
   name: string;
@@ -41,8 +31,11 @@ export type MenuListProp = {
   onMenuItemClick: (path: string) => void;
 };
 
-export type DraggablePostProp = {
-  post: object;
-  onTransitionRaise: (newDirection: TransitionDirection) => void;
-  drawerRef: RefObject<HTMLElement>;
+export type SideBarMenuProp = {
+  container: RefObject<HTMLElement>;
+  onBackClick: () => void;
+  onSettingClick: () => void;
+  onMenuItemClick: (path: string) => void;
+  overlayScope: AnimationScope<any>;
+  contentScope: AnimationScope<any>;
 };
