+++ src/Page/Home/Components/index.ts
@@ -1,0 +1,5 @@
+import { SideBarMenu } from './SidebarMenu/SidebarMenu';
+import { DraggablePost } from './DraggablePost/DraggablePost';
+import { CommentDrawer } from './CommentDrawer/CommentDrawer';
+
+export { SideBarMenu, DraggablePost, CommentDrawer };
