+++ src/Page/Home/HomePage.styled.ts
@@ -1,0 +1,18 @@
+import { styled } from '@/stitches.config';
+import { motion } from 'framer-motion';
+import * as Dialog from '@radix-ui/react-dialog';
+import * as Layout from '@Layouts/DefaultLayout';
+
+export const PageContainer = styled('div', {
+  position: 'relative',
+  display: 'flex',
+  flexDirection: 'column',
+  width: '100dvw',
+  height: '100dvh',
+});
+
+export const ContentSectionWrapper = styled('div', {
+  flex: 1,
+  position: 'relative',
+  // width:
+});
